feat: update select docs
